Projection pursuit classification tree with random variable selection in each split

Find tree structure using various projection pursuit indices of classification in each split.

Usage

PPtree_splitMOD(form, data, PPmethod = 'LDA', 
size.p = 1,  lambda = 0.1, entro , entroindiv,...)

Arguments

form

A character with the name of the class variable.

data

Data frame with the complete data set.

PPmethod

index to use for projection pursuit: 'LDA', 'PDA'

size.p

proportion of variables randomly sampled in each split, default is 1, returns a PPtree.

lambda

penalty parameter in PDA index and is between 0 to 1 . If lambda = 0, no penalty parameter is added and the PDA index is the same as LDA index. If lambda = 1 all variables are treated as uncorrelated. The default value is lambda = 0.1.

entro

TRUE, compute the entropy method

entroindiv

TRUE, compute the entropy for each observation in 1D projection

...

arguments to be passed to methods

Value

An object of class PPtreeclass with components

Tree.Struct

Tree structure of projection pursuit classification tree

projbest.node

1-dim optimal projections of each split node

splitCutoff.node

cutoff values of each split node

origclass

original class

origdata

original data

References

Lee, YD, Cook, D., Park JW, and Lee, EK (2013) PPtree: Projection pursuit classification tree, Electronic Journal of Statistics, 7:1369-1386.